requirements.txt
lunarcalendar==0.0.9
lark-oapi==1.2.9
python-dotenv==1.0.0
uvicorn==0.30.0
fastapi==0.111.0